+++ build_app.sh
@@ -66,8 +66,8 @@ cat > "$APP/Contents/Info.plist" <<PLIST
   <key>CFBundleExecutable</key>      <string>${EXEC_NAME}</string>
   <key>CFBundleIconFile</key>        <string>${EXEC_NAME}</string>
   <key>CFBundlePackageType</key>     <string>APPL</string>
-  <key>CFBundleVersion</key>         <string>1.2.0</string>
-  <key>CFBundleShortVersionString</key><string>1.2.0</string>
+  <key>CFBundleVersion</key>         <string>1.3.0</string>
+  <key>CFBundleShortVersionString</key><string>1.3.0</string>
   <key>LSMinimumSystemVersion</key>  <string>11.0</string>
   <key>NSHighResolutionCapable</key> <true/>
   <key>NSMicrophoneUsageDescription</key>